XD-1156 Refactor Plugins
- Created AbstractPlugin with no-op/common
implementing methods and have other concrete Plugins
extend this
- Removed unused methods in JobPlugin
- In StreamPlugin, use ChannelInterceptorAware type check when adding tap
to the producer
