+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/plugins/MBeanExportingPlugin.java
@@ -16,11 +16,9 @@ package org.springframework.xd.dirt.plugins;
 import java.util.Properties;
 
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.xd.dirt.container.XDContainer;
 import org.springframework.xd.module.core.Module;
-import org.springframework.xd.module.core.Plugin;
 
 /**
  * Exports MBeans from a module using a unique domain name xd.[group].[module]
@@ -28,7 +26,7 @@ import org.springframework.xd.module.core.Plugin;
  * @author David Turanski
  * @author Gary Russell
  */
-public class MBeanExportingPlugin implements Plugin {
+public class MBeanExportingPlugin extends AbstractPlugin {
 
 	private static final String CONTEXT_CONFIG_ROOT = XDContainer.XD_CONFIG_ROOT + "plugins/jmx/";
 
@@ -43,22 +41,6 @@ public class MBeanExportingPlugin implements Plugin {
 		objectNameProperties.put("xd.module.index", module.getDeploymentMetadata().getIndex());
 
 		module.addProperties(objectNameProperties);
-	}
-
-	@Override
-	public void postProcessModule(Module module) {
-	}
-
-	@Override
-	public void beforeShutdown(Module module) {
-	}
-
-	@Override
-	public void removeModule(Module module) {
-	}
-
-	@Override
-	public void preProcessSharedContext(ConfigurableApplicationContext context) {
 	}
 
 	@Override
