--- spring-xd-module/src/main/java/org/springframework/xd/module/core/PluginAdapter.java
@@ -1,53 +1,0 @@
-/*
- * Copyright 2013 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.xd.module.core;
-
-import org.springframework.context.ConfigurableApplicationContext;
-
-
-/**
- * 
- * @author David Turanski
- */
-public class PluginAdapter implements Plugin {
-
-	@Override
-	public void preProcessModule(Module module) {
-	}
-
-	@Override
-	public void postProcessModule(Module module) {
-	}
-
-	@Override
-	public void removeModule(Module module) {
-	}
-
-	@Override
-	public void beforeShutdown(Module module) {
-	}
-
-	@Override
-	public void preProcessSharedContext(ConfigurableApplicationContext context) {
-	}
-
-	@Override
-	public boolean supports(Module module) {
-		return true;
-	}
-
-}
